Remove sampler and add domain checking
The equation of motion for the jump process needs to have some checks on
the domain to ensure it doesn't crash. I also had to add the
DiffEqCallbacks package to call a function that resets the current size
of the time step to be small when the number of crosslinkers is changed.